reform test_layer2_API with googletest framwork and revise mkdir & mknod return type

+++ include/files.h
@@ -18,8 +18,8 @@ class FilesOperation {
     void unlink_inode(u_int64_t inode_number);
     u_int64_t disk_namei(const char* path);
     u_int64_t namei(const char* path);
-    u_int64_t fischl_mkdir(const char*, mode_t);
-    u_int64_t fischl_mknod(const char*, mode_t);
+    int fischl_mkdir(const char*, mode_t);
+    int fischl_mknod(const char*, mode_t);
     //int fischl_readdir(const char *, void *, fuse_fill_dir_t, off_t, struct fuse_file_info *, enum fuse_readdir_flags);
     int fischl_unlink (const char *);
     //int fischl_open (const char *, struct fuse_file_info *);
